Clean up unused files + update Form1.cs + update ImportSprites.cs + add SpriteDatabase.cs
+kas and Asar+ have been deleted, as they're pointless. Form1 has been
modified to reflect that. Importing (not finished) and sprite list
functions have been added to ImportSprites.cs. SpriteDatabase.cs is the
sprite list function, which simply lists all of the sprites located in
Podoboo's resources. It is not nearly as advanced as the rest of Podoboo
and is not completely finished.